Importing InterViews 3.2a tarball, dated 27 June 1994.  ftp://ftp.sgi.com/graphics/interviews/3.2a.tar.Z

+++ iv/src/include/IV-2_6/InterViews/button.h
@@ -49,7 +49,7 @@ public:
     void operator=(ButtonState&);
     void SetValue(int);
     void SetValue(void*);
-    void GetValue (int& v) { v = (int)value; }
+    void GetValue (int& v) { v = int(long(value)); }
     void GetValue (void*& v) { v = value; }
 protected:
     void* value;
